requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
gspread>=5.10.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pytz>=2023.3
plotly>=5.17.0
python-dotenv>=1.0.0
schedule>=1.2.0
APScheduler>=3.10.4 